+++ src/js/tabs.js
@@ -2,7 +2,7 @@
 // @deno-types=./43.ts
 import { $$, attr, css, makelogger, on } from "./19.js"
 import { internals, observeattributes, stylize, tag, validate } from "./43.js"
-import { FocusGroupMixin } from "./focusgroup.js"
+import { FocusGroupMixin } from "./focus.js"
 import { ariaRelatives, ariaState, AriaMultiSelectable, AriaSelected } from "./aria.js"
 
 const ilog = makelogger("tabs")
