+++ src/js/toolbar.js
@@ -1,6 +1,6 @@
 // @deno-types=./43.ts
 import { internals, tag } from "./43.js"
-import { FocusGroupMixin } from "./focusgroup.js"
+import { FocusGroupMixin } from "./focus.js"
 
 export const Toolbar = tag(
   "aria-toolbar",
